Refactor: Auth context
