--- src/types/AuthType.ts
@@ -1,11 +1,0 @@
-export interface User {
-  id: number;
-  name: string;
-  email: string;
-}
-
-export interface Auth {
-  user: User | null;
-  login: (email: string, password: string) => void;
-  logout: () => void;
-}
