+++ src/types/UserType.ts
@@ -1,0 +1,5 @@
+export interface User {
+  id: number;
+  name: string;
+  email: string;
+}
